requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv
psycopg2-binary
requests
pgvector
python-telegram-bot
aiohttp
pydantic